requirements.txt
Django>=5.0,<6.0
weasyprint>=62.0
django-ckeditor>=6.7.1
xhtml2pdf>=0.2.15
